Refactoring; split `tests/forms/test_user` test suite into ... * `tests/forms/test_user_register` * `tests/forms/test_user_administer.py` * `tests/forms/test_user_edit_details.py` and * `tests/forms/test_user_change_password.py`
