+++ tests/forms/test_client_edit.py
@@ -3,8 +3,6 @@
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from random import choice
-
 import pytest
 from flask_babel import gettext as _
 from wtforms.validators import ValidationError
